[config-postgis-refactor] Postgis-removing refactoring complete
