--- geoIpClient/src/main/scala/CpGeoConfig.scala
@@ -1,3 +1,0 @@
-package se.lu.nateko.cp.geoipclient
-
-case class CpGeoConfig(baseUri: String, maxAgeDays: Int, emailErrorsTo: String)
