+++ georestheart/src/main/resources/reference.conf
@@ -1,4 +1,4 @@
-geoIpClient{
+geoipclient{
 	baseUri = "https://geoip.icos-cp.eu"
 	maxAgeDays = 90
 	emailErrorsTo = "carbon.admin@nateko.lu.se"
