+++ src/main/resources/application.conf
@@ -104,8 +104,6 @@ cpauth {
 		skipInit = false
 	}
 
-	geoip = ${geoIpClient}
-
 	mailing = ${cpauthCore.mailing}
 	mailing.password = "dummy (replace in proper config)"
